change field placeholders to values
when the settings are shown in the fields as placeholder text, I have to re-type the entire text if I want to make a small change. I believe it is better to set the field's value property instead. Also, when the field contents are too long to fit within the field there is no way view the hidden parts of the placeholder text, but a field value can be scrolled to the end (to verify a long mp3 file path, for example).